% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_alignment.R
\name{create_alignment}
\alias{create_alignment}
\title{Create a D&D alignment meme}
\usage{
create_alignment(
  lawful_good,
  lawful_neutral,
  lawful_evil,
  neutral_good,
  true_neutral,
  neutral_evil,
  chaotic_good,
  chaotic_neutral,
  chaotic_evil
)
}
\arguments{
\item{lawful_good}{string with the image path}

\item{lawful_neutral}{string with the image path}

\item{lawful_evil}{string with the image path}

\item{neutral_good}{string with the image path}

\item{true_neutral}{string with the image path}

\item{neutral_evil}{string with the image path}

\item{chaotic_good}{string with the image path}

\item{chaotic_neutral}{string with the image path}

\item{chaotic_evil}{string with the image path}
}
\value{
an image
}
\description{
Create a D&D alignment meme
}
