% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_alignment.R
\name{create_alignment}
\alias{create_alignment}
\title{Create a D&D alignment meme}
\usage{
create_alignment(
  lawful_good = NULL,
  lawful_neutral = NULL,
  lawful_evil = NULL,
  neutral_good = NULL,
  true_neutral = NULL,
  neutral_evil = NULL,
  chaotic_good = NULL,
  chaotic_neutral = NULL,
  chaotic_evil = NULL,
  placeholder = NULL
)
}
\arguments{
\item{lawful_good}{string with the image path}

\item{lawful_neutral}{string with the image path}

\item{lawful_evil}{string with the image path}

\item{neutral_good}{string with the image path}

\item{true_neutral}{string with the image path}

\item{neutral_evil}{string with the image path}

\item{chaotic_good}{string with the image path}

\item{chaotic_neutral}{string with the image path}

\item{chaotic_evil}{string with the image path}

\item{placeholder}{string with the imagem path to a placeholder image
(not used alignment)}
}
\value{
an image
}
\description{
Create a D&D alignment meme
}
